requirements.txt
flask==2.3.3
python-dotenv==1.0.0
pillow==10.0.1
openai>=1.0.0